+++ src/router/index.jsx
@@ -3,6 +3,7 @@ import App from '../App'
 import Home from '../pages/Home'
 import ProductDetail from '../pages/ProductDetail'
 import Cart from '../pages/Cart'
+import Wishlist from '../pages/Wishlist'
 import Login from '../pages/Login'
 import Signup from '../pages/Signup'
 import Mypage from '../pages/Mypage'
@@ -17,6 +18,7 @@ const router = createBrowserRouter([
       { index: true, element: <Home /> },
       { path: 'product/:id', element: <ProductDetail /> },
       { path: 'cart', element: <Cart /> },
+      { path: 'wishlist', element: <Wishlist /> },
       { path: 'mypage', element: <PrivateRoute><Mypage /></PrivateRoute> },
     ],
   },